Regularly ensure the NAS is mounted and restart the dependencies if it wasn't
When the power fails and restarts (or when I stop the multiplug and turn it back on
again afterwards), the raspberry-pis start faster than the NAS, causing the rr services
and transmission to be missing the data.

We regularly run a script making sure that the NAS directories are mounted, and restart
rr/transmission if they were not.